+++ larpmanager/accounting/base.py
@@ -1,0 +1,32 @@
+# LarpManager - https://larpmanager.com
+# Copyright (C) 2025 Scanagatta Mauro
+#
+# This file is part of LarpManager and is dual-licensed:
+#
+# 1. Under the terms of the GNU Affero General Public License (AGPL) version 3,
+#    as published by the Free Software Foundation. You may use, modify, and
+#    distribute this file under those terms.
+#
+# 2. Under a commercial license, allowing use in closed-source or proprietary
+#    environments without the obligations of the AGPL.
+#
+# If you have obtained this file under the AGPL, and you make it available over
+# a network, you must also make the complete source code available under the same license.
+#
+# For more information or to purchase a commercial license, contact:
+# commercial@larpmanager.com
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later OR Proprietary
+
+from larpmanager.cache.feature import get_event_features
+
+
+def is_reg_provisional(instance, features=None):
+    if not features:
+        features = get_event_features(instance.run.event_id)
+
+    if "payment" in features:
+        if instance.tot_iscr > 0 >= instance.tot_payed:
+            return True
+
+    return False
