+++ larpmanager/templates/larpmanager/manage/orga.html
@@ -11,6 +11,29 @@
         - {{ run.get_development_display }}
         - <a href="{% url 'register' run.event.slug run.number %}">{{ registration_status }}</a>
         <hr />
+        <a href="{% url 'orga_registrations' run.event.slug run.number %}">{% trans "Registrations" %}</a>: {{ counts.count_reg }}
+        {% if counts.count_provisional %}
+            - ({% trans "provisional" %}: {{ counts.count_provisional }}
+        {% endif %}
+        {% if counts.count_staff %}
+            - ({% trans "staff" %}: {{ counts.count_staff }}
+        {% endif %}
+        {% if counts.count_fill %}
+            - ({% trans "filler" %}: {{ counts.count_fill }}
+        {% endif %}
+        {% if counts.count_wait %}
+            - ({% trans "waiting" %}: {{ counts.count_wait }}
+        {% endif %}
+        {% if counts.count_collaborator %}
+            - ({% trans "collaborator" %}: {{ counts.count_collaborator }}
+        {% endif %}
+        {% if counts.count_npc %}
+            - ({% trans "npc" %}: {{ counts.count_npc }}
+        {% endif %}
+        {% if counts.count_seller %}
+            - ({% trans "seller" %}: {{ counts.count_seller }}
+        {% endif %}
+        <hr />
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
